Add fix.

+++ frontend/src/components/SubscriptionReport.js
@@ -49,7 +49,7 @@ const SubscriptionReport = ({ timeRange, url, id, type }) => {
               <Text placeSelf="center">{metric.display_name}</Text>
               <Text p={2} fontSize="42px" placeSelf="center">
                 {metric.display_name == "Total weth earned"
-                  ? Number.parseInt(Number.parseInt(metric.value) / 1e18)
+                  ? Number.parseFloat(Number.parseFloat(metric.value) / 1e18)
                   : metric.value}
               </Text>
             </Flex>
